[BUGFIX] Fixed variable type

+++ remote_pe_reader.cpp
@@ -2,7 +2,7 @@
 
 bool read_module_header(HANDLE processHandle, BYTE *start_addr, size_t mod_size, OUT BYTE* buffer, const size_t buffer_size)
 {
-	DWORD read_size = 0;
+	SIZE_T read_size = 0;
 	ReadProcessMemory(processHandle, start_addr, buffer, buffer_size, &read_size);
 	if (get_nt_hrds(buffer) == NULL) {
 		printf("[-] Cannot get the module header!\n");
@@ -17,7 +17,7 @@ bool read_module_header(HANDLE processHandle, BYTE *start_addr, size_t mod_size,
 BYTE* get_module_section(HANDLE processHandle, BYTE *start_addr, size_t mod_size, const size_t section_num, OUT size_t &section_size)
 {
 	BYTE header_buffer[HEADER_SIZE] = { 0 };
-	DWORD read_size = 0;
+	SIZE_T read_size = 0;
 
 	if (!read_module_header(processHandle, start_addr, mod_size, header_buffer, HEADER_SIZE)) {
 		return NULL;
@@ -46,7 +46,7 @@ void free_module_section(BYTE *section_buffer)
 
 size_t read_pe_from_memory(const HANDLE processHandle, BYTE *start_addr, const size_t mod_size, OUT BYTE* buffer)
 {
-	DWORD read_size = 0;
+	SIZE_T read_size = 0;
 	if (ReadProcessMemory(processHandle, start_addr, buffer, mod_size, &read_size)) {
 		return read_size;
 	}
@@ -60,7 +60,7 @@ size_t read_pe_from_memory(const HANDLE processHandle, BYTE *start_addr, const s
 	//if not possible to read full module at once, try to read it section by section:
 	size_t sections_count = get_sections_count(hdr_buffer, HEADER_SIZE);
 	for (size_t i = 0; i < sections_count; i++) {
-		DWORD read_sec_size = 0;
+		SIZE_T read_sec_size = 0;
 		PIMAGE_SECTION_HEADER hdr = get_section_hdr(hdr_buffer, HEADER_SIZE, i);
 		if (!hdr) {
 			printf("[-] Failed to read the header of section: %d\n", i);
